Fix bundling issue with render-shared
Moving the renderResultFinalOutput to render-shared somehow introduced a bundling failure (out of order). This addresses that.
