+++ src/command/render/render.ts
@@ -107,6 +107,7 @@ import {
   RenderExecuteOptions,
   RenderFilesResult,
   RenderOptions,
+  RenderResult,
 } from "./types.ts";
 import {
   ExecuteResult,
@@ -576,6 +577,53 @@ export function removePandocTo(renderOptions: RenderOptions) {
     renderOptions.pandocArgs = removePandocToArg(renderOptions.pandocArgs);
   }
   return renderOptions;
+}
+
+export function renderResultFinalOutput(
+  renderResults: RenderResult,
+  relativeToInputDir?: string,
+) {
+  // final output defaults to the first output of the first result
+  let result = renderResults.files[0];
+  if (!result) {
+    return undefined;
+  }
+
+  // see if we can find an index.html instead
+  for (const fileResult of renderResults.files) {
+    if (fileResult.file === "index.html") {
+      result = fileResult;
+      break;
+    }
+  }
+
+  // determine final output
+  let finalInput = result.input;
+  let finalOutput = result.file;
+
+  if (renderResults.baseDir) {
+    finalInput = join(renderResults.baseDir, finalInput);
+    if (renderResults.outputDir) {
+      finalOutput = join(
+        renderResults.baseDir,
+        renderResults.outputDir,
+        finalOutput,
+      );
+    } else {
+      finalOutput = join(renderResults.baseDir, finalOutput);
+    }
+  } else {
+    finalOutput = join(dirname(finalInput), finalOutput);
+  }
+
+  // return a path relative to the input file
+  if (relativeToInputDir) {
+    const inputRealPath = Deno.realPathSync(relativeToInputDir);
+    const outputRealPath = Deno.realPathSync(finalOutput);
+    return relative(inputRealPath, outputRealPath);
+  } else {
+    return finalOutput;
+  }
 }
 
 // default pandoc renderer immediately renders each execute result
